Refactor state test imports to use AbstractFSM. Update TransitionFunction references in multiple state test files for improved consistency and clarity.
